Group the onboarding/settings helpers under the services that own them
Owner directive B (2026-08-06): the sparse top-level functions scattered
across the ij-plugin onboarding/settings cluster consolidate into their
owning services — devrig-common stays the library of pure functions; the
plugin side groups by owner, so "how does the IDE half install and find
devrig" has one address instead of five files of loose helpers.

Who moved where (grouping + consolidation only, no behavior change):

- DevrigSetupRunner.Companion now owns the whole setup surface:
  devrigInstalled (was its own file, DevrigInstallProbe.kt),
  devrigBinPath, devrigInstallFailedMarker, startDownloadPoller, and the
  installer-output parser (InstallerStep + parseInstallerLine, was
  InstallerProgress.kt). ProcessLineBuffer nests in the runner. The
  parser and probe are exactly as testable as before — companion
  members, pinned by the same tests.
- DevrigPromotion.Companion owns the promotion content: the registry
  key, devrigPromotionEnabled, and the balloon body renderer
  devrigInstallOfferBody.
- McpSteroidConfigurable.Companion owns the page's own helpers:
  copyWithFeedback + COPIED_HINT (the copy receipt), and
  devrigStdioMcpConfigJson — the page is that snippet's only consumer
  (the settings-page twin of 'devrig install config'), so it lives with
  the page, still delegating to DevrigUserLauncher.invocation.

Tests follow their subjects: DevrigInstallProbeTest and
InstallerProgressTest merge into DevrigSetupTest,
DevrigInstallOfferBodyTest into DevrigPromotionTest, and the stdio
snippet pin moves into McpSteroidConfigurableTest. The ij-plugin README
code map drops the two absorbed files.
