+++ ij-plugin/src/test/kotlin/com/jonnyzzz/mcpSteroid/onboarding/DevrigSetupTest.kt
@@ -4,9 +4,11 @@ package com.jonnyzzz.mcpSteroid.onboarding
 import com.intellij.openapi.application.ModalityState
 import com.intellij.openapi.progress.ProcessCanceledException
 import com.intellij.openapi.progress.ProgressIndicator
-import com.jonnyzzz.mcpSteroid.aiAgents.StdioMcpCommand
 import com.jonnyzzz.mcpSteroid.devrig.devrigLauncherDisplayPath
-import com.jonnyzzz.mcpSteroid.aiAgents.stdioMcpServersJson
+import com.jonnyzzz.mcpSteroid.onboarding.DevrigSetupRunner.Companion.devrigBinPath
+import com.jonnyzzz.mcpSteroid.onboarding.DevrigSetupRunner.Companion.devrigInstalled
+import com.jonnyzzz.mcpSteroid.onboarding.DevrigSetupRunner.Companion.parseInstallerLine
+import com.jonnyzzz.mcpSteroid.onboarding.DevrigSetupRunner.Companion.startDownloadPoller
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancel
@@ -15,9 +17,12 @@ import kotlinx.coroutines.isActive
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withTimeout
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Assert.fail
 import org.junit.Test
+import java.nio.file.Files
 import java.nio.file.Path
 import java.util.concurrent.atomic.AtomicBoolean
 import java.util.concurrent.atomic.AtomicInteger
@@ -36,6 +41,19 @@ class DevrigSetupTest {
         assertEquals(Path.of("/home/u/.mcp-steroid/bin/devrig.cmd"), devrigBinPath(home, windows = true))
     }
 
+    @Test
+    fun `devrigInstalled checks the per-OS launcher file`() {
+        val home = Files.createTempDirectory("home")
+        assertFalse(devrigInstalled(home, windows = false))
+        val bin = Files.createDirectories(home.resolve(".mcp-steroid").resolve("bin"))
+        Files.createFile(bin.resolve("devrig"))
+        assertTrue(devrigInstalled(home, windows = false))
+        // Windows looks for devrig.cmd, not devrig.
+        assertFalse(devrigInstalled(home, windows = true))
+        Files.createFile(bin.resolve("devrig.cmd"))
+        assertTrue(devrigInstalled(home, windows = true))
+    }
+
     /**
      * The launcher path the settings page RENDERS and the path this module CHECKS on disk are built by two
      * functions in two modules — they must name the same file, or the page shows a command that does not
@@ -47,31 +65,6 @@ class DevrigSetupTest {
         assertEquals(
             devrigBinPath(Path.of("/home/u"), windows = false).toString(),
             devrigLauncherDisplayPath("/home/u", windows = false),
-        )
-    }
-
-    /**
-     * What the settings page offers to paste into Cursor (or any other client devrig has no CLI for) must be
-     * the registration devrig itself writes — same launcher, same `mcp` subcommand — or the manual path
-     * silently stops matching the automatic one.
-     */
-    @Test
-    fun `the stdio snippet points at the stable launcher and matches what devrig registers`() {
-        val home = Path.of("/home/u")
-
-        val posix = devrigStdioMcpConfigJson(home, windows = false)
-        assertEquals(
-            stdioMcpServersJson(StdioMcpCommand("/home/u/.mcp-steroid/bin/devrig", listOf("mcp"))),
-            posix,
-        )
-
-        // The Windows case pins the cmd.exe wrapping and the quoting, not the path separator: this test
-        // also runs on POSIX, where Path.resolve joins with '/'. Separators are the bin-path test's job.
-        val winHome = Path.of("C:\\Users\\u")
-        val launcher = devrigBinPath(winHome, windows = true).toString()
-        assertEquals(
-            stdioMcpServersJson(StdioMcpCommand("cmd.exe", listOf("/d", "/c", "\"$launcher\" mcp"))),
-            devrigStdioMcpConfigJson(winHome, windows = true),
         )
     }
 
@@ -148,7 +141,7 @@ class DevrigSetupTest {
     @Test
     fun `line buffer reassembles chunks and emits only completed lines`() {
         val lines = mutableListOf<String>()
-        val buffer = ProcessLineBuffer { lines += it }
+        val buffer = DevrigSetupRunner.ProcessLineBuffer { lines += it }
         buffer.append("[mcp-steroid] down")
         buffer.append("loading devrig (~226 MB) ...\n[mcp-steroid] SHA")
         assertEquals(listOf("[mcp-steroid] downloading devrig (~226 MB) ..."), lines)
@@ -161,13 +154,13 @@ class DevrigSetupTest {
 
     /**
      * The installer's final line — the `ERROR: …` reason on the failures that matter most — can arrive
-     * with no trailing newline. Before [ProcessLineBuffer.flush] existed it was silently dropped and the
-     * user got the generic exit-code message instead of the reason.
+     * with no trailing newline. Before [DevrigSetupRunner.ProcessLineBuffer.flush] existed it was silently
+     * dropped and the user got the generic exit-code message instead of the reason.
      */
     @Test
     fun `flush emits the trailing ERROR line that has no newline`() {
         val lines = mutableListOf<String>()
-        val buffer = ProcessLineBuffer { lines += it }
+        val buffer = DevrigSetupRunner.ProcessLineBuffer { lines += it }
         buffer.append("[mcp-steroid] platform: macos-arm64\n[mcp-steroid] ERROR: no space left on device")
         assertEquals(listOf("[mcp-steroid] platform: macos-arm64"), lines)
 
@@ -180,6 +173,69 @@ class DevrigSetupTest {
 
         buffer.flush() // nothing left: flushing again must not re-emit
         assertEquals(2, lines.size)
+    }
+
+    // --- parseInstallerLine: the installer's output as a wire contract for the progress UI ---
+    // These are real lines from `installer-gen/src/main/resources/templates/install.sh.tmpl` (and the
+    // matching `install.ps1.tmpl` wording). If the templates change their phrasing, these tests fail
+    // instead of the progress bar silently going blank.
+
+    @Test
+    fun `downloading line yields the phase and the total size`() {
+        val step = parseInstallerLine("[mcp-steroid] downloading jdk (~385 MB) from https://example.com/jdk.tar.gz ...")
+        assertTrue(step!!.text, step.text.contains("Downloading jdk"))
+        assertTrue(step.text, step.text.contains("385 MB"))
+        assertEquals(385L * 1024 * 1024, step.totalBytes)
+        assertEquals(false, step.isError)
+    }
+
+    @Test
+    fun `devrig download is recognised too`() {
+        val step = parseInstallerLine("[mcp-steroid] downloading devrig (~226 MB) from https://example.com/devrig.zip ...")
+        assertEquals(226L * 1024 * 1024, step!!.totalBytes)
+        assertTrue(step.text, step.text.contains("devrig"))
+    }
+
+    @Test
+    fun `retry, verify, reuse, register and ready lines map to phases without a size`() {
+        val cases = mapOf(
+            "[mcp-steroid] attempt 2/3 failed (curl exited 28); retrying in 4s..." to "2/3",
+            "[mcp-steroid] SHA-256 verified: abc123" to "Verifying",
+            "[mcp-steroid] already installed: devrig-macos-arm64-0.101-abc123def456" to "reusing",
+            "[mcp-steroid] another install finished first; using existing tree" to "reusing",
+            "[mcp-steroid] registering devrig (devrig install devrig)..." to "Registering",
+            "[mcp-steroid] devrig binary is ready." to "installed",
+            "[mcp-steroid] platform: macos-arm64" to "macos-arm64",
+        )
+        for ((line, expected) in cases) {
+            val step = parseInstallerLine(line)
+            assertTrue("$line -> ${step?.text}", step != null && step.text.contains(expected, ignoreCase = true))
+            assertNull("$line must not carry a size", step!!.totalBytes)
+            assertEquals(line, false, step.isError)
+        }
+    }
+
+    @Test
+    fun `the installer's own ERROR line is surfaced as the failure reason`() {
+        val step = parseInstallerLine("[mcp-steroid] ERROR: insufficient disk space in /home/u/.mcp-steroid/binaries: need ~1800 MB")
+        assertTrue(step!!.isError)
+        assertTrue(step.text, step.text.startsWith("insufficient disk space"))
+    }
+
+    @Test
+    fun `lines that are not ours, or carry no step, are ignored`() {
+        // curl's progress bar, shell noise, and the installer's help text must not overwrite the phase.
+        assertNull(parseInstallerLine("######################################                     54.2%"))
+        assertNull(parseInstallerLine("bash: line 3: warning: something"))
+        assertNull(parseInstallerLine("[mcp-steroid] "))
+        assertNull(parseInstallerLine("[mcp-steroid]     Debian/Ubuntu:  sudo apt-get install -y curl unzip tar"))
+        assertNull(parseInstallerLine(""))
+    }
+
+    @Test
+    fun `leading and trailing whitespace does not hide a step`() {
+        val step = parseInstallerLine("  [mcp-steroid] registering devrig (devrig install devrig)...  \r")
+        assertTrue(step?.text ?: "null", step != null && step.text.contains("Registering"))
     }
 
     private fun waitUntil(what: String, condition: () -> Boolean) {
